Changed session and scan serializers to work more smoothly with the FE.
